requirements/contact_requirements.txt
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-rest-swagger==2.2.0
djangogrpcframework==0.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
grpcio==1.31.0
grpcio-tools==1.31.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.2.0
pkg-resources==0.0.0
postgis==1.0.4
protobuf==3.13.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.10
